requirements.txt
streamlit
paramiko
python-dotenv
pandas
pytest
pytest-mock
google-generativeai
requests

# LangChain dependencies
langchain
langchain-core
langchain-google-genai
langchain-anthropic
langchain-community

# LangGraph multi-agent system
langgraph
langgraph-supervisor
